% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state
A funtion that maps the location of the accident during a specific year in one or more states.
Returns error if state number is not present or year is incorrect.
If there are no accidents (hopefully) it returns a message 'No accident to plot'}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A character or numeric containing the number state to  map (coerced to integer in the function).}

\item{year}{A character or numeric element containing the year to summirise.}
}
\value{
returns map with the accident as points.
}
\description{
fars_map_state
A funtion that maps the location of the accident during a specific year in one or more states.
Returns error if state number is not present or year is incorrect.
If there are no accidents (hopefully) it returns a message 'No accident to plot'
}
\examples{
fars_map_state(1,2015)
fars_map_state('20','2014')

}
