% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read}
\alias{fars_read}
\title{fars_read
A funtion that checks if the file exists and returns the content as tbl_df.
No need to be exported as it's mainly used inside the package}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character containing the filename built with the \code{make_filename} function}
}
\value{
returns a dataframe (more exactly a tbl_df from dplyr) with the information from
the US National Highway Traffic Safety Administration's Fatality Analysis Reporting System.
If the filename doesn't exists it returs an error.
}
\description{
fars_read
A funtion that checks if the file exists and returns the content as tbl_df.
No need to be exported as it's mainly used inside the package
}
\examples{
\dontrun{
fars_read('accident_2015.csv.bz2')
fars_read(make_filename(2015))
}

}
