% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years
A funtion that create a list of dataset combining the  \code{make_filename}  \code{fars_read} functions.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A character or numeric vector containing the years to import.}
}
\value{
returns a list of dataframes (more exactly a tbl_df from dplyr) with the information from
the US National Highway Traffic Safety Administration's Fatality Analysis Reporting System.
If the filename with the reference year doesn't exists it returs a warning and a NULL element in the list.
}
\description{
fars_read_years
A funtion that create a list of dataset combining the  \code{make_filename}  \code{fars_read} functions.
}
\examples{
fars_read_years(c(2014,2015))
fars_read_years(c('2013','2014'))

}
