% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years
A funtion that summarise how many accident per year and month are registered in the FARS dataset.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A character or numeric vector containing the years to summirise.}
}
\value{
returns a dataframe (more exactly a tbl_df from dplyr) with the number of accident per Year/Month
in the US National Highway Traffic Safety Administration's Fatality Analysis Reporting System.
It will return a warning if the years are not presented in the dataset.
}
\description{
fars_summarize_years
A funtion that summarise how many accident per year and month are registered in the FARS dataset.
}
\examples{
fars_summarize_years(c(2014,2015))
fars_summarize_years(c('2013','2014'))

}
