% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{make_filename}
\alias{make_filename}
\title{make_filename
A function that returns the filename for the FARS database given a certain year.
No need to be exported as it's mainly used inside the package}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A character or number (it will be coerced to integer in the function) containing the year}
}
\value{
returns a character with the correct file name to be searched in the FARS database.
It will return an error if input is incorrect (ie '2015c')
}
\description{
make_filename
A function that returns the filename for the FARS database given a certain year.
No need to be exported as it's mainly used inside the package
}
\examples{
\dontrun{
make_filename(2015)
make_filename('2015')
}

}
