% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{For a given state and given year, it draws the location of accidents}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{an integer representing state number for which the location will be
drawn}

\item{year}{a vector of years for which FARS data file will be loaded}
}
\value{
if there is no accident a massage is produced. for invalid state number error
 is thrown. for valid input, plots a map with accident as dots on it.
}
\description{
For a given state and given year, it draws the location of accidents
}
\examples{
\dontrun{
fars_map_state (3, 2013)
fars_map_state (10, 2015)
}

}
