% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Draws observations as point on a map}
\usage{
fars_map_state(state.num, year, internal = FALSE)
}
\arguments{
\item{state.num}{a State number, an integer between 1 and 56.}

\item{year}{a year, as integer.}

\item{internal}{logical, if the function is run using data from the package,
defaults to FALSE}
}
\value{
\code{NULL (empty)} returned invisibly \code{\link{invisible}}.
  As side effect, the function will plot a State map with the observations
  (accidents) plotted as points on the map.
}
\description{
\code{fars_map_state} takes a State number and a year as arguments.
  Reads the datafile for the corresponding year, and draws a map of the
  State with points at the longitude and latitude of the observations /
  accidents in the dataset.
}
\details{
If the value for the \code{state.num} parameter does not correspond to a
  valid State in the dataset, the function will throw an error.
  If the State and year combination has zero observations, a message is
  printed.
  All the points outside of Longitude > 900 or Latitude > 90 are nullified and
  not drawn on the map.
}
\note{
The data files must be in the current working directory (see also
  \code{\link{getwd}}) to be found.
}
\section{Warning}{

  \pkg{maps} package should be attached (\code{library(maps)}),
  or the function will throw an error because maps does not export the data
  (\code{object 'stateMapEnv'}) for the \code{\link{map}} function.
}

\examples{
\donttest{fars_map_state(1, 2013)}
\donttest{fars_map_state(48, 2015)}

}
