% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string.}
}
\value{
This function returns a data frame from reading \code{filename}.
}
\description{
Generates a data frame from a file. If the file does not exist
the function will stop with an error message.
}
\examples{
fars_read("accident_2014.csv.bz2")
fars_read("accident_2015.csv.bz2")
}

