% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads a data file and returns a tibble}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{path to a csv data file, as a character string.}
}
\value{
A tibble, enhanced implementation of data.frame
  (see \code{\link{tibble}}).
}
\description{
Reads a csv file and import the data, returning a tibble.
}
\details{
The data type in the columns is guessed automatically by the
  \code{\link[readr]{read_csv}} function in the \pkg{readr} package.
  Additional messages and progress are supressed to limit the console
  output.

  If the file is not found at the specified path, the function throws
  an error.
}
\examples{
\donttest{data2013 <- fars_read('fars_data/accident_2013.csv.bz2')}
\donttest{data2014 <- fars_read('fars_data/accident_2014.csv.bz2')}

}
