% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of numbers representing years for analysis}
}
\value{
This function returns a list of data.frames with the
   length of the vector \code{years} and two columns: MONTH and year.
}
\description{
Reads files into a list of data.frames with two columns, MONTH and year,
over the vector of years input. All values in the vector years need to be
convertable into an integer otherwise the make_filename function wll error.
}
\examples{
fars_read_years(c(2014,2015))
fars_read_years(2013:2015)
}

