% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads a vector of files and returns months and years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of years, as integers.}
}
\value{
A list of tibbles (see \code{\link{tibble}})).
}
\description{
The function reads a list of csv data files, specified by year, and returns
  a list of tibbles \code{\link{tibble}} with two columns: month number and
  year combinations.
}
\details{
If any of the years leads to an inexistent file, the function
  throws a warning for each.
}
\note{
The data files must be in the current working directory (see also
  \code{\link{getwd}}) to be found.
}
\section{Warning}{

  \code{\link{dplyr}} package should be attached (\code{library(dplyr)}),
  or the function will throw an error because it cannot find the pipe
  operator.
}

\examples{
year_list <- fars_read_years(2013)
year_list <- fars_read_years(c(2013, 2014, 2015))

}
\seealso{
(\code{\link{make_filename}})
}
