% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Monthly number of accident for given years in FARS data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector of years for which FARS data file will be loaded and number
of accident will be summurized}
}
\value{
return NULL for an invalid year. For a valid input, returns list of
FARS data in data frame in years and number of accidents
}
\description{
Monthly number of accident for given years in FARS data
}
\examples{
\dontrun{
fars_summarize_years(c(2001, 2005, 2007))
fars_summarize_years(c("2001", "2005", "2007"))
}

}
