% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of numbers representing years for analysis}
}
\value{
This function returns a data.frame with months as rows
   and years as columns. Each combination contains the count of
   accidents.
}
\description{
Summarizes the count of accidents by month and year for the input
vector \code{years}.
}
\examples{
\dontrun{
fars_summarize_years(2012:2014)
}
}

