% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Reads a vector of files and returns total observations by month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector of years, as integers.}
}
\value{
A tibble, enhanced implementation of data.frame
  (see \code{\link{tibble}}), with 12 rows and 1 column for each year.
}
\description{
The function takes a vector of year numbers, reads the corresponding files
  (using \code{\link{fars_read_years}}) and counts the number of rows in
  each, aggregated by month number.
  Then it builds a single table with one row per month number, and a column
  for each file / year, where the values represent the total number of
  observations.
}
\note{
The data files must be in the current working directory (see also
  \code{\link{getwd}}) to be found.
}
\examples{
\donttest{aggregated <- fars_summarize_years(2013)}
\donttest{aggregated <- fars_summarize_years(c(2013, 2014, 2015))}

}
\seealso{
(\code{\link{fars_read_years}})
}
