% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Builds a filename as character string}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a year number, as integer.}
}
\value{
A filename as a character string.
}
\description{
Inserts a year number into a base string, in order to build a
  filename.
}
\details{
The resulting string can be passed to other functions to process
  the target data file. It is a convenience function to process many files
  that differ only in the year number.
}
\examples{
print(make_filename(2013))
filename <- make_filename(2014)

}
\seealso{
(\code{\link{fars_read}})
}
